Split engine options into general and network-related.
