+++ api.go
@@ -11,40 +11,9 @@ import (
 	"google.golang.org/grpc"
 )
 
-// DefaultListenAddress is the default TCP address for the gRPC listener.
-const DefaultListenAddress = ":50555"
-
-// WithListenAddress returns an option that sets the TCP address for the gRPC
-// listener.
-//
-// If this option is omitted or addr is empty DefaultListenAddress is used.
-func WithListenAddress(addr string) EngineOption {
-	if addr != "" {
-		_, port, err := net.SplitHostPort(addr)
-		if err != nil {
-			panic(fmt.Sprintf("invalid listen address: %s", err))
-		}
-
-		if _, err := net.LookupPort("tcp", port); err != nil {
-			panic(fmt.Sprintf("invalid listen address: %s", err))
-		}
-	}
-
-	return func(opts *engineOptions) {
-		opts.ListenAddress = addr
-	}
-}
-
-// WithServerOptions returns an option that adds gRPC server options.
-func WithServerOptions(options ...grpc.ServerOption) EngineOption {
-	return func(opts *engineOptions) {
-		opts.ServerOptions = append(opts.ServerOptions, options...)
-	}
-}
-
 // serveAPI runs the gRPC server.
 func serveAPI(ctx context.Context, opts *engineOptions) error {
-	s := grpc.NewServer(opts.ServerOptions...)
+	s := grpc.NewServer(opts.Network.ServerOptions...)
 
 	// config []RichApplication to []Application
 	configs := make([]configkit.Application, len(opts.AppConfigs))
@@ -53,7 +22,7 @@ func serveAPI(ctx context.Context, opts *engineOptions) error {
 	}
 	api.RegisterServer(s, configs...)
 
-	lis, err := net.Listen("tcp", opts.ListenAddress)
+	lis, err := net.Listen("tcp", opts.Network.ListenAddress)
 	if err != nil {
 		return fmt.Errorf("unable to start gRPC listener: %w", err)
 	}
@@ -74,7 +43,7 @@ func serveAPI(ctx context.Context, opts *engineOptions) error {
 	logging.Log(
 		opts.Logger,
 		"listening for API requests on %s",
-		opts.ListenAddress,
+		opts.Network.ListenAddress,
 	)
 
 	err = s.Serve(lis)
