+++ app.go
@@ -2,32 +2,10 @@ package infix
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dodeca/logging"
-	"github.com/dogmatiq/dogma"
 )
-
-// WithApplication returns an option that hosts an additional application on the
-// engine.
-func WithApplication(app dogma.Application) EngineOption {
-	return func(opts *engineOptions) {
-		cfg := configkit.FromApplication(app)
-
-		for _, c := range opts.AppConfigs {
-			if c.Identity().ConflictsWith(cfg.Identity()) {
-				panic(fmt.Sprintf(
-					"can not host both %s and %s because they have conflicting identities",
-					c.Identity(),
-					cfg.Identity(),
-				))
-			}
-		}
-
-		opts.AppConfigs = append(opts.AppConfigs, cfg)
-	}
-}
 
 func hostApplication(
 	ctx context.Context,
