+++ discovery.go
@@ -3,66 +3,11 @@ package infix
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/configkit/api/discovery"
 	"github.com/dogmatiq/dodeca/logging"
-	"github.com/dogmatiq/linger"
-	"github.com/dogmatiq/linger/backoff"
 )
-
-// Discoverer is a function that notifies an observer when a config API target
-// becomes available or unavailable.
-//
-// It blocks until ctx is canceled or a fatal error occurs.
-type Discoverer func(ctx context.Context, o discovery.TargetObserver) error
-
-// WithDiscoverer returns an option that sets the discoverer used to find other
-// engine instances.
-//
-// If this option is omitted or d is nil, no discovery is performed.
-//
-// TODO: Use Bonjour as the default discovery mechanism.
-// See https://github.com/dogmatiq/configkit/issues/58.
-func WithDiscoverer(d Discoverer) EngineOption {
-	return func(opts *engineOptions) {
-		opts.Discoverer = d
-	}
-}
-
-// DefaultDialer is the default dialer used to connect to other engine
-// instances.
-var DefaultDialer = discovery.DefaultDialer
-
-// WithDialer returns an option that sets the dialer used to connect to other
-// engine instances.
-//
-// If this option is omitted or d is nil, DefaultDialer is used.
-//
-// This option must be used in conjunction with WithDiscoverer().
-func WithDialer(d discovery.Dialer) EngineOption {
-	return func(opts *engineOptions) {
-		opts.Dialer = d
-	}
-}
-
-// DefaultDialerBackoffStrategy is the default backoff strategy for the dialer.
-var DefaultDialerBackoffStrategy backoff.Strategy = backoff.WithTransforms(
-	backoff.Exponential(100*time.Millisecond),
-	linger.FullJitter,
-	linger.Limiter(0, 30*time.Second),
-)
-
-// WithDialerBackoffStrategy returns an option that sets the strategy used to
-// determine when the engine should retry dialing another engine instance.
-//
-// If this option is omitted or s is nil DefaultDialerBackoffStrategy is used.
-func WithDialerBackoffStrategy(s backoff.Strategy) EngineOption {
-	return func(opts *engineOptions) {
-		opts.DialerBackoffStrategy = s
-	}
-}
 
 // discover runs the API server discovery system, if configured.
 func discover(
@@ -70,10 +15,6 @@ func discover(
 	opts *engineOptions,
 	obs discovery.ApplicationObserver,
 ) error {
-	if opts.Discoverer == nil {
-		return nil
-	}
-
 	i := &discovery.Inspector{
 		Observer: discovery.NewApplicationObserverSet(
 			obs,
@@ -115,12 +56,12 @@ func discover(
 				i.Run(ctx, c)
 			},
 		},
-		Dial:            opts.Dialer,
-		BackoffStrategy: opts.DialerBackoffStrategy,
+		Dial:            opts.Network.Dialer,
+		BackoffStrategy: opts.Network.DialerBackoff,
 		Logger:          opts.Logger,
 	}
 
-	err := opts.Discoverer(
+	err := opts.Network.Discoverer(
 		ctx,
 		&discovery.TargetExecutor{
 			Task: func(ctx context.Context, t *discovery.Target) {
